requirements.txt
requests==2.32.4
beautifulsoup4==4.13.4
google-generativeai==0.8.5
python-dotenv==1.1.1
